fix: tooltip
